Feature : 포인트 계산 함수 분리
